+++ src/api/scope/index.js
@@ -9,7 +9,9 @@ import buildInScope from './lib/build-in-scope';
 import testInScope from './lib/test-in-scope';
 import modifyCIProps from './lib/modify-ci-props';
 import ciUpdateAction from './lib/ci-update-action';
-const scopeConfig = require('./lib/scope-config');
+import scopeConfig from './lib/scope-config';
+import catScope from './lib/cat-scope';
+import refreshScope from './lib/refresh-scope';
 
 export {
   catObject,
@@ -24,4 +26,6 @@ export {
   modifyCIProps,
   ciUpdateAction,
   scopeConfig,
+  catScope,
+  refreshScope,
 };